update to sdk 2.3

+++ ISV/App/App.cpp
@@ -102,7 +102,7 @@ uint32_t g_cert_size = 0;
 uint8_t *gvc_ias_res = NULL, *gvc_ias_sig = NULL, *gvc_ias_crt = NULL;
 uint32_t gvc_ias_res_size, gvc_ias_sig_size, gvc_ias_crt_size;
 uint8_t *priv_rl = NULL, *sig_rl = NULL;
-size_t priv_rl_size, sig_rl_size;
+uint32_t priv_rl_size, sig_rl_size;
 uint8_t cur_ts[AS_TS_SIZE + 1];
 
 void print_array(uint8_t* array, uint32_t array_size, bool debug = false) {
@@ -142,7 +142,7 @@ void print_array(uint8_t* array, uint32_t array_size, bool debug = false) {
 uint32_t as_get_quote(
     sgx_report_t &report,
     uint8_t *&p_quote,
-    size_t &quote_size,
+    uint32_t &quote_size,
     const char* service_domain)
 {
     uint32_t res = -1;
@@ -192,7 +192,7 @@ void as_attestation(
     sgx_status_t ret = SGX_ERROR_UNEXPECTED;
     uint32_t enclave_ret = -1;
     uint8_t *as_quote = NULL;
-    size_t as_quote_size = 0;
+    uint32_t as_quote_size = 0;
 
     do
     {
